renaming of masterRoot to bip44Root
