+++ src/bip44RootToAccountPrivKey.js
